Resolve "Support bfs benchmark"

Closes #154

See merge request akita/gcn3!138

Former-commit-id: c5c316d2d5de0425cdda11bd323bf23d8a38d2cc
Former-commit-id: 397a02739e07302d52a9b84e66500655398a1fa5
Former-commit-id: 4b975ce3f17f93e4a926e564256d0085fd10f540

+++ emu/aluvopc.go
@@ -198,7 +198,7 @@ func (u *ALUImpl) runVCmpEqU32(state InstEmuState) {
 			continue
 		}
 
-		if sp.SRC0[i] == sp.SRC1[i] {
+		if uint32(sp.SRC0[i]) == uint32(sp.SRC1[i]) {
 			sp.VCC = sp.VCC | (1 << i)
 		}
 	}
